use uint64_t for timer period.

+++ rcl/test/rcl/test_timer.cpp
@@ -789,7 +789,7 @@ TEST_F(TestPreInitTimer, test_timer_clock) {
 TEST_F(TestPreInitTimer, test_timer_call) {
   int64_t next_call_start = 0;
   int64_t next_call_end = 0;
-  int64_t old_period = 0;
+  uint64_t old_period = 0;
   times_called = 0;
 
   EXPECT_EQ(RCL_RET_OK, rcl_timer_get_time_until_next_call(&timer, &next_call_start));
@@ -804,7 +804,7 @@ TEST_F(TestPreInitTimer, test_timer_call) {
 
   next_call_start = next_call_end;
   ASSERT_EQ(RCL_RET_OK, rcl_timer_exchange_period(&timer, 0, &old_period));
-  EXPECT_EQ(RCL_S_TO_NS(1), old_period);
+  EXPECT_EQ((uint64_t)RCL_S_TO_NS(1), old_period);
   ASSERT_EQ(RCL_RET_OK, rcl_timer_call(&timer)) << rcl_get_error_string().str;
   EXPECT_EQ(times_called, 4);
   EXPECT_EQ(RCL_RET_OK, rcl_timer_get_time_until_next_call(&timer, &next_call_end));
@@ -884,9 +884,9 @@ TEST_F(TestPreInitTimer, test_invalid_init_fini) {
 }
 
 TEST_F(TestPreInitTimer, test_timer_get_period) {
-  int64_t period = 0;
+  uint64_t period = 0;
   ASSERT_EQ(RCL_RET_OK, rcl_timer_get_period(&timer, &period));
-  EXPECT_EQ(RCL_S_TO_NS(1), period);
+  EXPECT_EQ((uint64_t)RCL_S_TO_NS(1), period);
 
   EXPECT_EQ(RCL_RET_INVALID_ARGUMENT, rcl_timer_get_period(nullptr, &period));
   rcl_reset_error();
